cm/MFA-Conformer-infer/pyproject.toml
[project]
name = "mfa-conformer-infer"
version = "1.0.0"
description = "MFA-Conformer inference-only service for deepfake audio detection"
requires-python = "==3.10.12"
dependencies = [
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "nemo-toolkit[asr]>=2.5.3",
    "speechbrain>=1.0.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "torchinfo>=1.8.0",
    "numpy",
    "tqdm",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
